Moving Metadata specific downstream route extensions to the Metadata folder
