+++ src/Ocelot/Metadata/DownstreamRouteExtensions.cs
@@ -1,8 +1,9 @@
-﻿using System.Globalization;
+﻿using Ocelot.Configuration;
+using System.Globalization;
 using System.Reflection;
 using System.Text.Json;
 
-namespace Ocelot.Configuration;
+namespace Ocelot.Metadata;
 
 public static class DownstreamRouteExtensions
 {
