+++ test/Ocelot.UnitTests/Configuration/DownstreamRouteExtensionsTests.cs
@@ -1,6 +1,7 @@
 ﻿using Ocelot.Configuration;
 using Ocelot.Configuration.Creator;
 using Ocelot.Configuration.File;
+using Ocelot.Metadata;
 using Ocelot.Values;
 using System.Numerics;
 using System.Text.Json;
